build($release): config release webpack
