+++ antd-tools/generator-types/index.js
@@ -7,9 +7,9 @@ try {
   parseAndWrite({
     version: pkg.version,
     name: 'types',
-    path: path.resolve(rootPath, './v2-doc/src/docs'),
+    path: path.resolve(rootPath, './examples/docs'),
     // default match lang
-    test: /en-US\.md/,
+    test: /zh-CN\.md/,
     outputDir: path.resolve(rootPath, './vetur'),
     tagPrefix: 'a-',
   });
