+++ components/loading/src/size.ts
