+++ components/modal-pro/src/components/ModalHeader.vue
@@ -5,21 +5,5 @@
     {{ title }}
   </BasicTitle>
 </template>
-<script lang="ts">
-  import { defineComponent } from 'vue';
-
-  import { BasicTitle } from '../../../basic-title';
-  import useConfigInject from '../../../_util/hooks/useConfigInject';
-  import { basicProps } from '../props';
-
-  export default defineComponent({
-    name: 'AModalProHeader',
-    components: { BasicTitle },
-    props: basicProps,
-    setup(props) {
-      const { prefixCls: prefixClsNew } = useConfigInject('modal-pro-title', props);
-      return { prefixClsNew };
-    },
-  });
-</script>
+<script lang="ts" src="./modal-header.ts"></script>
