+++ components/modal-pro/src/components/ModalWrapper.vue
@@ -12,4 +12,5 @@
     </a-spin>
   </ContainerScroll>
 </template>
+
 <script lang="ts" src="./modal-wrapper.ts"></script>
