+++ components/select-api/src/empty.tsx
@@ -1,0 +1,5 @@
+import PRESENTED_IMAGE_SIMPLE from '../../empty/simple';
+
+const simpleEmptyImg = <PRESENTED_IMAGE_SIMPLE />;
+
+export default simpleEmptyImg;
