--- examples/utils/env.ts
@@ -1,3 +1,0 @@
-export function isSite(): boolean {
-  return import.meta.env.VITE_APP_SITE === 'true';
-}
